// source/resources/examples/typescript/exchanges/interfaces/ExchangeSymbolConfig.ts
/*
 * Copyright 2019-present NEM Foundation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { Address, MosaicId, NamespaceId, NetworkType } from 'symbol-sdk';

export interface ExchangeSymbolConfig {
  apiUrl: string;
  networkType: NetworkType;
  networkGenerationHashSeed: string;
  centralAccountAddress: Address;
  tokenId: MosaicId;
  tokenAlias: NamespaceId;
  tokenDivisibility: number;
  requiredConfirmations: number;
  useFinalization: boolean;
}